% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadablePlotly.R
\name{downloadablePlotly}
\alias{downloadablePlotly}
\title{Server-side logic for plotly with download data button in modebar}
\usage{
downloadablePlotly(
  input,
  output,
  session,
  plot,
  filename,
  content,
  title = "Download plot data"
)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate.}

\item{plot}{Object of class \code{'plotly'} or a function or reactive that generates one.}

\item{filename}{A string of the filename, including extension, that the
user's web browser should default to when downloading the file; or a
function that returns such a string. (Reactive values and functions may be
used from this function.)}

\item{content}{A function that takes a single argument \code{file} that is a
file path (string) of a nonexistent temp file, and writes the content to
that file path. (Reactive values and functions may be used from this
function.)}

\item{title}{Text for plotly tooltip.}
}
\value{
No return value, called to generate server logic.
}
\description{
Server-side logic for plotly with download data button in modebar
}
\seealso{
\code{\link{downloadablePlotlyUI}}
}
