% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenDownload.R
\name{hiddenDownloadUI}
\alias{hiddenDownloadUI}
\title{UI for hidden download link}
\usage{
hiddenDownloadUI(id)
}
\arguments{
\item{id}{id string to be namespaced.}
}
\description{
UI for hidden download link
}
\seealso{
\code{\link[shiny]{NS}}, \code{\link{hiddenDownload}}
}
\keyword{internal}
