% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadablePlot.R
\name{downloadablePlotUI}
\alias{downloadablePlotUI}
\title{UI for plot with download data button}
\usage{
downloadablePlotUI(
  id,
  title = "Download plot data",
  width = "100\%",
  height = "400px"
)
}
\arguments{
\item{id}{id string that gets namespaced by \code{shiny::NS}.}

\item{title}{Text to display on hover of download button.}

\item{...}{Named arguments passed to \code{plotOutput}.}
}
\description{
UI for plot with download data button
}
\seealso{
\code{\link[shiny]{NS}}, \code{\link{downloadablePlot}}, \code{\link{plotOutput}}
}
